se3313a-jchin55-yzha2229-swei57-ggupta24-master/requirements.txt
backcall==0.1.0
decorator==4.3.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pygame==1.9.4
Pygments==2.3.0
python-chess==0.23.10
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
